Split out the query string preparation code.
