+++ lib/processImage.js
@@ -6,6 +6,8 @@ const mime = require('mime');
 const accepts = require('accepts');
 const hijackResponse = require('hijackresponse');
 
+const prepareImproQueryString = require('./prepareImproQueryString');
+
 const isImageByExtension = {};
 Object.keys(mime._extensions).forEach(contentType => {
   if (/^image\//.test(contentType)) {
@@ -16,63 +18,8 @@ Object.keys(mime._extensions).forEach(contentType => {
 
 isImageByExtension.jpg = true;
 
-const isSupportedEngine = {
-  gm: true,
-  sharp: true,
-  pngcrush: true,
-  pngquant: true,
-  jpegtran: true,
-  optipng: true,
-  svgfilter: true,
-  inkscape: true
-};
-
-const queryStringEngineAndArgsRegex = new RegExp(
-  `^(${Object.keys(isSupportedEngine).join('|')})=(.*)`
-);
-
 function isImageExtension(extension) {
   return isImageByExtension[extension.toLowerCase()];
-}
-
-function prepareQueryString(queryString) {
-  let m;
-
-  if ((m = queryString.match(queryStringEngineAndArgsRegex)) !== null) {
-    const [, engineName, engineArgs] = m;
-    const result = [engineName];
-    const remaining = engineArgs.split(',');
-
-    let isEngineOptions = false;
-    let engineOptions;
-
-    remaining.forEach(bit => {
-      if (engineName === 'svgfilter' && bit[0] === '-' && bit[1] === '-') {
-        if (!isEngineOptions) {
-          isEngineOptions = true;
-          engineOptions = [];
-        }
-        engineOptions.push(bit.slice(2));
-      } else if (bit[0] === '-') {
-        result.push(bit.slice(1));
-      } else if (bit === 'horizontal' || bit === 'vertical') {
-        const indexOfArg = result.indexOf('flip');
-        if (indexOfArg > -1) {
-          result[indexOfArg] += `=${bit}`;
-        } else {
-          // XXX
-        }
-      }
-    });
-
-    if (isEngineOptions) {
-      result[0] += `=${engineOptions.join('+')}`;
-    }
-
-    return result.join('&');
-  }
-
-  return queryString;
 }
 
 function reverseIteratorFor(arr) {
@@ -164,7 +111,7 @@ module.exports = options => {
           let sourceMetadata;
 
           function makeFilterInfosAndTargetFormat() {
-            const preparedQueryString = prepareQueryString(queryString);
+            const preparedQueryString = prepareImproQueryString(queryString);
             const parseResult = impro.parse(
               preparedQueryString,
               _.defaults(
